Fix profile and user tests

+++ src/routes/profiles/index.ts
@@ -5,7 +5,8 @@ import type { ProfileEntity } from '../../utils/DB/entities/DBProfiles';
 
 enum Message {
   PROFILE_NOT_FOUND = 'Profile not found',
-  PROFILE_FAKE_ID = 'Fake id param'
+  PROFILE_FAKE_ID = 'Fake id param',
+  PROFILE_EXISTS = 'User already has a profile'
 }
 
 const plugin: FastifyPluginAsyncJsonSchemaToTs = async (
@@ -46,6 +47,24 @@ const plugin: FastifyPluginAsyncJsonSchemaToTs = async (
       },
     },
     async function (request, reply): Promise<ProfileEntity> {
+      const profile = await fastify.db.profiles.findOne({
+        key: 'userId',
+        equals: request.body.userId
+      });
+
+      if (profile) {
+        throw reply.badRequest(Message.PROFILE_EXISTS);
+      }
+
+      const memberType = await fastify.db.profiles.findOne({
+        key: 'id',
+        equals: request.body.memberTypeId
+      });
+
+      if (!memberType) {
+        throw reply.badRequest(Message.PROFILE_FAKE_ID);
+      }
+
       return fastify.db.profiles.create(request.body);
     }
   );
@@ -58,13 +77,16 @@ const plugin: FastifyPluginAsyncJsonSchemaToTs = async (
       },
     },
     async function (request, reply): Promise<ProfileEntity> {
-      const deletedProfile = await fastify.db.profiles.delete(request.params.id);
+      const profile = await fastify.db.profiles.findOne({
+        key: 'id',
+        equals: request.params.id
+      });
 
-      if (!deletedProfile) {
-        throw reply.notFound(Message.PROFILE_NOT_FOUND);
+      if (!profile) {
+        throw reply.badRequest(Message.PROFILE_NOT_FOUND);
       }
 
-      return deletedProfile;
+      return fastify.db.profiles.delete(request.params.id);
     }
   );
